fix: simplify AI cleanup migrations and restore guest spec
Keep Change105 to AI people deletion and experimental-flag scrubbing,
split data vs table-drop into separate migrations, revert the already-
implemented guest-accounts spec, and regenerate the CLI catalog twice.
